Fixed a bug in subgeometries.gi. More precisely in the command IsFrameOfProjectiveSpace.  In order to verify the input there was a check put in place: coll := Collected(List(list,x->AmbientGeometry(x))); This gives an error since "list" is a list of projective spaces. This is due to the fact that Collected relies on Sort and GAP obviously doesn't know how to sort projective spaces. (Maybe the "Collected" command got changed in a recent GAP update??). I fixed this by replacing Collected by DuplicateFreeList. This does not use the Sort command and seems to work fine.
I've documented this in detail here, because this might give rise to more bugs in FinInG. As general rule, it is better to use DuplicateFree instead of Collected, and keep in mind that Collected also Sorts the elements. I don't know if there is also a "non-sorting" version of Collected.